User Login.

+++ base/views.py
@@ -1,10 +1,50 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
+from django.contrib import messages
 from .models import Room, Topic
 from .forms import RoomForm
 
 
 # Create your views here.
+
+def loginPage(request):
+    page = 'login'
+
+    if request.user.is_authenticated:
+        return redirect('home')
+
+    if request.method =='POST':
+        username = request.POST.get('username')
+        password = request.POST.get('password')
+
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, 'Username or Password is incorrect')
+        user = authenticate(request ,username=username, password=password)
+
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Username or Password is incorrect')
+        
+    context = {'page': page}
+    return render(request, 'base/login_register.html', context)
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+def registerPage(request):
+    form = UserCreationForm()
+    return render(request, 'base/login_register.html', {'form': form})
+
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
     rooms = Room.objects.filter(
@@ -29,7 +69,7 @@ def room(request, pk):
 
     return render(request, 'base/room.html', context)
 
-
+@login_required(login_url='/login/')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -43,9 +83,13 @@ def createRoom(request):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='/login/')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -59,8 +103,14 @@ def updateRoom(request, pk):
     return render(request, 'base/room_form.html', context)
 
 
-def delateRoom(request, pk):
+@login_required(login_url='/login/')
+def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!')
+
+
     if request.method == "POST":
         room.delete()
         return redirect('home')
